+++ events/interactionCreate.js
@@ -1,4 +1,3 @@
-import { getFoodItemArray } from '../lib/tools.js';
 const once = false;
 const name = 'interactionCreate';
 
@@ -9,20 +8,7 @@ async function invoke(interaction) {
 		(await import(`#commands/${interaction.commandName}`)).invoke(interaction);
 	}
 	if (interaction.isAutocomplete()){
-		if (interaction.commandName === "recipe") { 
-			const focusedOption = interaction.options.getFocused(true);
-			let choices = getFoodItemArray()
-			const filtered = choices.filter(choice => choice.name.toLowerCase().includes(focusedOption.value));
-			let options;
-			if (filtered.length > 25) {
-				options = filtered.slice(0, 25);
-			} else {
-				options = filtered;
-			}
-			await interaction.respond(
-				options,
-			);
-		}
+		(await import(`#commands/${interaction.commandName}`)).autocomplete(interaction);
 	}
 }
